+++ app/Unit/Unit/Unit.jsx
@@ -18,18 +18,121 @@ class Unit extends React.Component {
   constructor(props){
     super(props);
     this.state = {
+      axios: false,
       close: false,
       onSpanBack: false
     };
+    this.axiosSource = axios.CancelToken.source();
     this._close_modal_Unit = this._close_modal_Unit.bind(this);
+    this._axios_getUnitImg = this._axios_getUnitImg.bind(this);
+    this._axios_getUnitData = this._axios_getUnitData.bind(this);
+    this._axios_get_UnitMount = this._axios_get_UnitMount.bind(this);
     this._handleEnter_spanBack = this._handleEnter_spanBack.bind(this);
     this._handleLeave_spanBack = this._handleLeave_spanBack.bind(this);
+    this._reset_UnitMount = ()=>{this._axios_get_UnitMount();};
     this.style={
 
     };
     //And! we have to 'hide' the scroll bar and preventing the scroll behavior to the page one for all
     //so dismiss the scroll ability for <body> here
     document.getElementsByTagName("BODY")[0].setAttribute("style","overflow-y:hidden;");
+  }
+
+  _axios_getUnitData(){
+    return axios.get('/router/units/'+this.unitId, {
+      headers: {
+        'charset': 'utf-8',
+        'token': window.localStorage['token']
+      }
+    })
+  };
+
+  _axios_getUnitImg(){
+    const self = this,
+          _axios_getUnitImg_base64 = (src)=>{
+            return axios.get('/router/img/'+src+'?type=unitSingle', {
+              headers: {
+                'token': window.localStorage['token']
+              }
+            });
+          };
+
+    return axios.get('/router/units/'+this.unitId+'/src', {
+      headers: {
+        'token': window.localStorage['token']
+      },
+      cancelToken: self.axiosSource.token
+    }).then((res)=>{
+      let resObj = JSON.parse(res.data);
+      let srcCover = resObj.main['pic_layer0'],
+          srcBeneath = resObj.main['pic_layer1'];
+
+      return axios.all([
+        _axios_getUnitImg_base64(srcCover),
+        srcBeneath? _axios_getUnitImg_base64(srcBeneath) : Promise.resolve({data: null})
+      ]).then(
+        axios.spread((resImgCover, resImgBeneath)=>{
+          let imgsBase64 = {
+            cover: resImgCover.data,
+            beneath: resImgBeneath.data
+          }
+          return imgsBase64;
+        })
+      )
+    }).catch(function (thrown) {
+      throw thrown;
+    });
+  };
+
+
+  _axios_get_UnitMount(){
+    const self = this;
+    let axiosArr = [this._axios_getUnitData(),this._axios_getUnitImg()];
+    this.setState({axios: true});
+
+    axios.all(axiosArr).then(
+      axios.spread(function(unitRes, imgsBase64){
+        self.setState({axios: false});
+        let resObj = JSON.parse(unitRes.data);
+        //we compose the marksset here, but sould consider done @ server
+        let keysArr = Object.keys(resObj.main.marksObj);//if any modified or update, keep the "key" as string
+        let [coverMarks, beneathMarks] = [{list:[],data:{}}, {list:[],data:{}}];
+        keysArr.forEach(function(key, index){
+          if(resObj.main.marksObj[key].layer==0){
+            coverMarks.data[key]=resObj.main.marksObj[key];
+            coverMarks.list[resObj.main.marksObj[key].serial] = key; //let the list based on order of marks, same as beneath
+          }else{
+            beneathMarks.data[key]=resObj.main.marksObj[key]
+            beneathMarks.list[resObj.main.marksObj[key].serial] = key;
+          }
+        })
+        //actually, beneath part might need to be rewritten to asure the state could stay consistency
+        self.props._set_store_UnitCurrent({
+          unitId:self.unitId,
+          identity: resObj.main.identity,
+          authorBasic: resObj.main.authorBasic,
+          coverSrc: imgsBase64.cover,
+          beneathSrc: imgsBase64.beneath,
+          coverMarksList:coverMarks.list,
+          coverMarksData:coverMarks.data,
+          beneathMarksList:beneathMarks.list,
+          beneathMarksData:beneathMarks.data,
+          nouns: resObj.main.nouns,
+          marksInteraction: resObj.main.marksInteraction,
+          broad: false,
+          refsArr: resObj.main.refsArr,
+          createdAt: resObj.main.createdAt
+        });
+      })
+    ).catch(function (thrown) {
+      if (axios.isCancel(thrown)) {
+        console.log('Request canceled: ', thrown.message);
+      } else {
+        console.log(thrown);
+        self.setState({axios: false});
+        alert("Failed, please try again later");
+      }
+    });
   }
 
   _handleEnter_spanBack(e){
@@ -61,16 +164,30 @@ class Unit extends React.Component {
   }
 
   componentDidUpdate(prevProps, prevState, snapshot){
-    //put in render() because we would jump from Unit to Unit without remount component
-    if(this.unitId !== prevProps.match.params.id) this.unitId = this.props.match.params.id;
+    //becuase there is chance we jump to another Unit from Related but using the same component
+    //so we check if the unit has changed
+    //but Notice! always check the diff between the current & pre id from 'path search'
+    //due to this is the only reliable and stable source (compare to the unitCurrent)
+    let prevParams = new URLSearchParams(prevProps.location.search); //we need value in URL query
+    if(this.unitId !== prevParams.get('unitId')){
+      //reset UnitCurrent to clear the view
+      //and Don't worry about the order between state reset, due to the Redux would keep always synchronized
+      let unitCurrentState = Object.assign({}, unitCurrentInit);
+      this.props._set_store_UnitCurrent(unitCurrentState);
+      this._axios_get_UnitMount();
+    };
   }
 
   componentDidMount(){
-    //set the current Unit id to this here to let the var could be modified
-    this.unitId = this.props.match.params.id;
+    //because we fetch the data of Unit only from this file,
+    //now we need to check if it was necessary to fetch or not in case the props.unitCurrent has already saved the right data we want
+    this._axios_get_UnitMount();
   }
 
   componentWillUnmount(){
+    if(this.state.axios){
+      this.axiosSource.cancel("component will unmount.")
+    }
     //reset UnitCurrent before leaving
     // It's Important !! next Unit should not have a 'coverSrc' to prevent children component render in UnitModal before Unit data response!
     let unitCurrentState = Object.assign({}, unitCurrentInit);
@@ -85,6 +202,7 @@ class Unit extends React.Component {
 
     let params = new URLSearchParams(this.props.location.search); //we need value in URL query
     let paramsTheater = params.has('theater'); //bool, true if there is 'theater'
+    this.unitId = params.get('unitId');
 
     return(
       <ModalBox containerId="root">
@@ -93,6 +211,7 @@ class Unit extends React.Component {
             (paramsTheater) ? (
               <Theater
                 {...this.props}
+                _reset_UnitMount={this._reset_UnitMount}
                 _close_theaterHeigher={this._close_modal_Unit}/>
             ): (
               <div
