+++ public/css/stylesSelfFront.css
@@ -111,24 +111,6 @@
   width: 100%;
   height: 91px;
 }
-._C_C_app_Unit_RelatedList_styles_module__boxList {
-  display: inline-block;
-  width: 73%;
-}
-
-._C_C_app_Unit_RelatedList_styles_module__boxSubtitle {
-  display: inline-block;
-  width: 27%;
-  box-sizing: border-box;
-  padding: 2% 1%;
-}
-
-._C_C_app_Unit_RelatedList_styles_module__fontSubtitle {
-  font-size: 1.44rem;
-  letter-spacing: 0.12rem;
-  font-weight: 400;
-  color: #777777;
-}
 ._C_C_app_Unit_RelatedOrigin_styles_module__comRelatedOrig {
   width: 100%;
   height: 100%;
@@ -185,6 +167,24 @@
   line-height: 1.77rem;
   word-wrap: break-word;
   color: #FAFAFA;
+}
+._C_C_app_Unit_RelatedList_styles_module__boxList {
+  display: inline-block;
+  width: 73%;
+}
+
+._C_C_app_Unit_RelatedList_styles_module__boxSubtitle {
+  display: inline-block;
+  width: 27%;
+  box-sizing: border-box;
+  padding: 2% 1%;
+}
+
+._C_C_app_Unit_RelatedList_styles_module__fontSubtitle {
+  font-size: 1.44rem;
+  letter-spacing: 0.12rem;
+  font-weight: 400;
+  color: #777777;
 }
 ._C_C_app_Unit_UnitLayerSwitch_styles_module__boxUnitLayerSwitch {
   width: 100%;
@@ -296,25 +296,6 @@
   line-height: 1.7rem;
   color: #ABABAB;
 }
-._C_C_app_Component_Blocks_SimpleBlock_styles_module__frameNail {
-  display: inline-block;
-  width: 32.5%;
-  min-width: 264px;
-  height: 19rem;
-  position: relative;
-  box-sizing: border-box;
-  margin: 0 0 18px;
-  box-shadow: 0 1px 1px -0.1rem, 0rem -0.05rem 2px -0.2rem;
-  border-radius: 0.3rem;
-  overflow: hidden
-}
-
-/* beneath must after the main 'frameNail'*/
-
-._C_C_app_Component_Blocks_SimpleBlock_styles_module__frameTwo {
-  width: 49.2%;
-  height: 19.4rem;
-}
 ._C_C_app_Component_Nails_NailBasic_styleNailBasic_module__Nails_Basic {
   background-color: #FFFFFF;
 }
@@ -340,6 +321,25 @@
   font-weight: 400;
   letter-spacing: 0.1rem;
   color: #757575;
+}
+._C_C_app_Component_Blocks_SimpleBlock_styles_module__frameNail {
+  display: inline-block;
+  width: 32.5%;
+  min-width: 264px;
+  height: 19rem;
+  position: relative;
+  box-sizing: border-box;
+  margin: 0 0 18px;
+  box-shadow: 0 1px 1px -0.1rem, 0rem -0.05rem 2px -0.2rem;
+  border-radius: 0.3rem;
+  overflow: hidden
+}
+
+/* beneath must after the main 'frameNail'*/
+
+._C_C_app_Component_Blocks_SimpleBlock_styles_module__frameTwo {
+  width: 49.2%;
+  height: 19.4rem;
 }
 ._C_C_app_Component_OpenedMark_styles_module__boxMarkBlock {
   width: 29%;
@@ -395,40 +395,6 @@
   word-break: break-all;
   color: #000000;
 }
-._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxInteraction {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  width: 100%;
-  min-height: 8.1rem;
-  position: sticky;
-  bottom: 0;
-  box-sizing: border-box;
-  background-color: #FCFCFC;
-}
-
-._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxPanelJump {
-  box-sizing: border-box;
-  padding: 0 10% 0 8%;
-  margin: 2.7vh 0 0 0;
-  text-align: right;
-}
-
-._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxContentDraft {
-  min-height: 37vh;
-  box-sizing: border-box;
-  padding: 1vh 9% 0 10.8%;
-}
-
-._C_C_app_Unit_Mark_AuthorBlock_styles_module__fontContentDraft {
-  font-size: 1.52rem;
-  letter-spacing: 0.07rem;
-  line-height: 1.6rem;
-  font-weight: 400;
-  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
-  word-break: break-all;
-  color: #000000;
-}
 ._C_C_app_Unit_Mark_ViewerBlock_styles_module__boxInteraction {
   display: flex;
   flex-direction: column;
@@ -455,6 +421,40 @@
 }
 
 ._C_C_app_Unit_Mark_ViewerBlock_styles_module__fontContentDraft {
+  font-size: 1.52rem;
+  letter-spacing: 0.07rem;
+  line-height: 1.6rem;
+  font-weight: 400;
+  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
+  word-break: break-all;
+  color: #000000;
+}
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxInteraction {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  width: 100%;
+  min-height: 8.1rem;
+  position: sticky;
+  bottom: 0;
+  box-sizing: border-box;
+  background-color: #FCFCFC;
+}
+
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxPanelJump {
+  box-sizing: border-box;
+  padding: 0 10% 0 8%;
+  margin: 2.7vh 0 0 0;
+  text-align: right;
+}
+
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxContentDraft {
+  min-height: 37vh;
+  box-sizing: border-box;
+  padding: 1vh 9% 0 10.8%;
+}
+
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__fontContentDraft {
   font-size: 1.52rem;
   letter-spacing: 0.07rem;
   line-height: 1.6rem;
